// web/api/cogniboard.schemas.ts
/**
 * Generated by orval v7.5.0 🍺
 * Do not edit manually.
 * CogniBoard
 * OpenAPI spec version: 0.0.1
 */
export interface AssignTaskDTO {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /**
   * Name of the person to assign the task to
   * @minLength 1
   */
  assignee_name: string;
}

export interface ChangeTaskStatusDTO {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /**
   * New status for the task
   * @minLength 1
   */
  status: string;
}

export interface CreateTaskDTO {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** Task's asignee (if any) */
  assignee_name?: string;
  /** Task's description */
  description?: string;
  /** Task's due date (if any) */
  due_date?: string;
  /**
   * Task's name
   * @minLength 1
   * @maxLength 50
   */
  title: string;
}

export interface ErrorDetail {
  /** Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id' */
  location?: string;
  /** Error message text */
  message?: string;
  /** The value at the given location */
  value?: unknown;
}

/**
 * Optional list of individual error details
 */
export type ErrorModelErrors = ErrorDetail[] | null;

export interface ErrorModel {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** A human-readable explanation specific to this occurrence of the problem. */
  detail?: string;
  /** Optional list of individual error details */
  errors?: ErrorModelErrors;
  /** A URI reference that identifies the specific occurrence of the problem. */
  instance?: string;
  /** HTTP status code */
  status?: number;
  /** A short, human-readable summary of the problem type. This value should not change between occurrences of the error. */
  title?: string;
  /** A URI reference to human-readable documentation for the error. */
  type?: string;
}

export type GetTasksDTOAssignee = string | null;

export type GetTasksDTOCompletedAt = string | null;

export type GetTasksDTODescription = string | null;

export type GetTasksDTODueDate = string | null;

export interface GetTasksDTO {
  assignee: GetTasksDTOAssignee;
  completed_at: GetTasksDTOCompletedAt;
  created_at: string;
  description: GetTasksDTODescription;
  due_date: GetTasksDTODueDate;
  id: string;
  status: string;
  title: string;
}

export type TasksTasks = GetTasksDTO[] | null;

export interface Tasks {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  tasks: TasksTasks;
}